Refactor Dockerfiles and scripts: streamline application startup, add database connection testing, and enhance database initialization process
