+++ backend/Dockerfile
@@ -55,5 +55,5 @@ EXPOSE 8000
 HEALTHCHECK --interval=30s --timeout=30s --start-period=5s --retries=3 \
     CMD curl -f http://localhost:8000/health || exit 1
 
-# Run the application
-CMD ["sh", "-c", "python scripts/init_db.py && python -m uvicorn main:app --host 0.0.0.0 --port 8000"]
+# Run the application for debugging
+CMD ["python", "scripts/init_db.py"]
