+++ backend/Dockerfile.prod
@@ -22,10 +22,16 @@ COPY backend/ .
 # Copy scripts directory for database initialization  
 COPY scripts/ ./scripts/
 
-
+# Create a startup script that initializes DB and then runs the app
+COPY scripts/start-production.sh ./start-production.sh
 
 # Create a non-root user
 RUN groupadd -r peluprice && useradd -r -g peluprice peluprice
+
+# Make scripts executable
+RUN chmod +x ./start-production.sh ./scripts/test_db_connection.py ./scripts/init_db.py
+
+# Change ownership
 RUN chown -R peluprice:peluprice /app
 
 # Set environment variables
@@ -42,5 +48,5 @@ EXPOSE 8000
 HEALTHCHECK --interval=30s --timeout=30s --start-period=5s --retries=3 \
     CMD curl -f http://localhost:8000/health || exit 1
 
-# Run the application
-CMD ["sh", "-c", "python scripts/init_db.py && python -m uvicorn main:app --host 0.0.0.0 --port 8000"]
+# Run the startup script
+CMD ["./start-production.sh"]
