+++ backend/app/models/__init__.py
@@ -1,2 +1,3 @@
-from .user import User, Base
+from ..database import Base
+from .user import User
 from .device import Device, DeviceStatus
