+++ backend/app/models/user.py
@@ -1,10 +1,8 @@
 
 from sqlalchemy import Column, Integer, String, DateTime
-from sqlalchemy.ext.declarative import declarative_base
 from sqlalchemy.orm import relationship
 from datetime import datetime
-
-Base = declarative_base()
+from ..database import Base
 
 class User(Base):
     __tablename__ = "users"
